Implement Encode/Decode on all transaction objects and turn the transaction version into an enum

+++ src/primitives/druid.rs
@@ -1,8 +1,9 @@
+use bincode::{Decode, Encode};
 use crate::primitives::asset::Asset;
 use serde::{Deserialize, Serialize};
 
 /// The expectation to be met in a specific DRUID transaction
-#[derive(Default, Clone, Debug, Ord, Eq, PartialEq, Serialize, Deserialize, PartialOrd)]
+#[derive(Default, Clone, Debug, Ord, PartialOrd, Eq, PartialEq, Serialize, Deserialize, Encode, Decode)]
 pub struct DruidExpectation {
     pub from: String,
     pub to: String,
@@ -16,10 +17,10 @@ pub struct DruidExpectation {
 /// `expect_value`          - The value expected by another party for this tx
 /// `expect_value_amount`   - The amount of the asset expected by another party for this tx
 /// `expect_address`        - The address the other party is expected to pay to
-#[derive(Default, Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
+#[derive(Default, Clone, Debug, Eq, PartialEq, Serialize, Deserialize, Encode, Decode)]
 pub struct DdeValues {
     pub druid: String,
-    pub participants: usize,
+    pub participants: u64,
     pub expectations: Vec<DruidExpectation>,
     pub genesis_hash: Option<String>,
 }
